DirectB2SManager

+++ src/main/java/de/mephisto/vpin/b2s/B2SThumbnailExtractor.java
@@ -1,5 +1,7 @@
 package de.mephisto.vpin.b2s;
 
+import de.mephisto.vpin.GameInfo;
+import de.mephisto.vpin.VPinServiceException;
 import de.mephisto.vpin.util.SystemInfo;
 import org.apache.commons.io.FilenameUtils;
 import org.apache.commons.io.IOUtils;
@@ -18,30 +20,34 @@ public class B2SThumbnailExtractor extends DefaultHandler {
   private final static Logger LOG = LoggerFactory.getLogger(B2SThumbnailExtractor.class);
 
   private String imageData;
+  private GameInfo game;
 
-  public B2SThumbnailExtractor() {
-    File vpxTablesFolder = SystemInfo.getInstance().getVPXTablesFolder();
-    File[] files = vpxTablesFolder.listFiles((dir, name) -> name.endsWith(".directb2s"));
-    for (File file : files) {
-      LOG.info("Extracting image from " + file.getAbsolutePath());
-      String name = FilenameUtils.getBaseName(file.getName());
-      File target = new File("./" , name + ".png");
-      parseImage(file, target);
-    }
+  public B2SThumbnailExtractor(GameInfo game) {
+    this.game = game;
   }
 
-  private File parseImage(File file, File target) {
+  public File extractImage(File file) throws VPinServiceException {
     try {
-      SAXParserFactory factory = SAXParserFactory.newInstance();
-      SAXParser saxParser = factory.newSAXParser();
-      saxParser.parse(file.getAbsolutePath(), this);
+      if (file.exists()) {
+        SAXParserFactory factory = SAXParserFactory.newInstance();
+        SAXParser saxParser = factory.newSAXParser();
+        saxParser.parse(file.getAbsolutePath(), this);
 
-      byte[] bytes = DatatypeConverter.parseBase64Binary(imageData);
-      FileOutputStream out = new FileOutputStream(target);
-      IOUtils.write(bytes, out);
-      out.close();
+        String baseName = FilenameUtils.getBaseName(game.getGameFileName());
+        File target = new File(SystemInfo.getInstance().getB2SImageExtractionFolder(), baseName + ".png");
+
+        byte[] bytes = DatatypeConverter.parseBase64Binary(imageData);
+        FileOutputStream out = new FileOutputStream(target);
+        IOUtils.write(bytes, out);
+        out.close();
+        LOG.info("Written uncropped directb2s image " + target.getAbsolutePath());
+
+        return target;
+      }
     } catch (Exception e) {
-      LOG.error("Failed to parse directb2s file '" + file.getAbsolutePath() + "': " + e.getMessage(), e);
+      String msg = "Failed to parse directb2s file '" + file.getAbsolutePath() + "': " + e.getMessage();
+      LOG.error(msg, e);
+      throw new VPinServiceException(msg, e);
     }
     return null;
   }
@@ -54,9 +60,5 @@ public class B2SThumbnailExtractor extends DefaultHandler {
         break;
       }
     }
-  }
-
-  public static void main(String args[]) {
-    new B2SThumbnailExtractor();
   }
 }
